# spec/openapi.json
{
  "components": {
    "schemas": {
...
      "PreferencesRequest": {
        "description": "Request payload for updating user preferences.",
        "properties": {
          "expectedRevision": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "interestThemeIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "type": [
              "array",
              "null"
            ]
          },
          "safetyToggleIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "type": [
              "array",
              "null"
            ]
          },
          "unitSystem": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "UserPreferencesResponse": {
        "description": "Response payload for user preferences.",
        "properties": {
          "interestThemeIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "type": "array"
          },
          "revision": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "safetyToggleIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "type": "array"
          },
          "unitSystem": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "interestThemeIds",
          "safetyToggleIds",
          "unitSystem",
          "revision",
          "updatedAt"
        ],
        "type": "object"
      },
      "crate.domain.Error": {
        "description": "OpenAPI schema for [`crate::domain::Error`].\n\nAPI error response payload with machine-readable code and human-readable\nmessage.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/crate.domain.ErrorCode",
            "description": "Stable machine-readable error code."
          },
          "details": {
            "description": "Supplementary error details for clients."
          },
          "message": {
            "description": "Human-readable message returned to clients.",
            "example": "Something went wrong",
            "type": "string"
          },
          "trace_id": {
            "description": "Correlation identifier for tracing this error across systems.",
            "example": "01HZY8B2W6X5Y7Z9ABCD1234",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "crate.domain.ErrorCode": {
        "description": "OpenAPI schema for [`crate::domain::ErrorCode`].\n\nStable machine-readable error codes returned in API error responses.",
        "enum": [
          "invalid_request",
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict",
          "service_unavailable",
          "internal_error"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Session cookie issued by POST /api/v1/login.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "HTTP interface for session-authenticated access and health probes.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "Wildside backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/me/preferences": {
      "get": {
        "description": "Fetch preferences, creating defaults if none exist.",
        "operationId": "getUserPreferences",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesResponse"
                }
              }
            },
            "description": "User preferences",
            "headers": {
              "Cache-Control": {
                "description": "Cache control header",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Fetch the authenticated user's preferences.",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "updateUserPreferences",
        "parameters": [
          {
            "description": "UUID for idempotent requests",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPreferencesResponse"
                }
              }
            },
            "description": "Updated preferences"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "summary": "Update the authenticated user's preferences.",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "servers": [
    {
      "description": "Relative to the deployment base URL",
      "url": "/"
    }
  ],
...
}